dbtexp/pyproject.toml
[tool.poetry]
name = "dbtexp"
version = "0.1.0"
description = ""
authors = ["Rajiv Mehta <rajiv.mehta@live.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
dbt-snowflake = "^1.4.1"
snowflake-connector-python = "2.7.11"
pandas = "^1.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
